refactor: create exception package
