+++ nextstep-jdbc/src/main/java/nextstep/jdbc/exception/SelectQueryFailException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class SelectQueryFailException extends RuntimeException {
     public SelectQueryFailException() {
