+++ nextstep-jdbc/src/main/java/nextstep/jdbc/exception/UpdateQueryFailException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 public class UpdateQueryFailException extends RuntimeException {
     public UpdateQueryFailException() {
